% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nested.R
\name{make_nested}
\alias{make_nested}
\title{Nested dataset to use downstream}
\usage{
make_nested(long_dataset, Gene, value, expression_level)
}
\arguments{
\item{long_dataset}{This is the dataset provided}

\item{Gene}{This is the list of genes.}

\item{value}{The is the value variable.}

\item{expression_level}{This is the expression level variable.}
}
\value{
nested_outcome
}
\description{
Creates nested table of grouped genes that may be useful for further modelling.
}
\examples{
data("data_two")
}
\author{
Miss Oriade Latifah Simpson, \email{s172084@dtu.dk}
}
