% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_data_one.R
\docType{data}
\name{my_data_one}
\alias{my_data_one}
\title{my_data_one The named dataset.}
\format{
A dataframe
}
\source{
https://github.com/ramhiser/datamicroarray/blob/master/data/west.RData
}
\usage{
data(my_data_one)
}
\description{
Data from Predicting the clinical status of human breast cancer by using gene expression profiles
}
\examples{
data(my_data_one)
}
\references{
Predicting the clinical status of human breast cancer by using gene expression profiles DOI: 10.1073/pnas.201162998
}
\keyword{datasets}
