% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_grand_pca.R
\docType{data}
\name{results_grand_pca}
\alias{results_grand_pca}
\title{Example dataset results_grand_pca  A PCA table}
\format{
A dataframe
}
\source{
https://github.com/ramhiser/datamicroarray/blob/master/data/west.RData
}
\usage{
data(results_grand_pca)
}
\description{
Data from Predicting the clinical status of human breast cancer by using gene expression profiles
}
\examples{
data(results_grand_pca)
}
\references{
Predicting the clinical status of human breast cancer by using gene expression profiles DOI: 10.1073/pnas.201162998
}
\keyword{datasets}
