% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_kmean_aug.R
\docType{data}
\name{results_kmean_aug}
\alias{results_kmean_aug}
\title{results_kmean_aug The augmented K maens results}
\format{
A dataframe
}
\source{
https://github.com/ramhiser/datamicroarray/blob/master/data/west.RData
}
\usage{
data(results_kmean_aug)
}
\description{
Data from Predicting the clinical status of human breast cancer by using gene expression profiles
}
\examples{
data(results_kmean_aug)
}
\references{
Predicting the clinical status of human breast cancer by using gene expression profiles DOI: 10.1073/pnas.201162998
}
\keyword{datasets}
