% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_pca_eig.R
\docType{data}
\name{results_pca_eig}
\alias{results_pca_eig}
\title{Example dataset PCA Eigenvalues}
\format{
A dataframe
}
\source{
https://github.com/ramhiser/datamicroarray/blob/master/data/west.RData
}
\usage{
data(results_pca_eig)
}
\description{
Data from Predicting the clinical status of human breast cancer by using gene expression profiles
}
\examples{
data(results_pca_eig)
}
\references{
Predicting the clinical status of human breast cancer by using gene expression profiles DOI: 10.1073/pnas.201162998
}
\keyword{datasets}
