% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{score_reverse}
\alias{score_reverse}
\alias{score_reverse.default}
\alias{score_reverse.data.frame}
\title{Reverse a score}
\usage{
score_reverse(x, range = NULL, ...)

\method{score_reverse}{default}(x, range = NULL, ...)

\method{score_reverse}{data.frame}(x, range = NULL, cols = NULL, ...)
}
\arguments{
\item{x}{A vector of \code{numeric} values}

\item{range}{A range of potential values}

\item{...}{Additional arguments passed to methods}

\item{cols}{An optional selection for columns}
}
\description{
Reverse a score
}
\examples{
x <- c(2L, 4L, 5L, 2L, 4L, 3L, 5L, 5L, 2L, 5L)
score_reverse(x)
score_reverse(x, 0:5)

df <- data.frame(x = x, y = score_reverse(x))
df

score_reverse(df)
score_reverse(df, range = 1:5)
score_reverse(df, range = list(0:10, 1:5))
score_reverse(df, range = list(y = 0:10, x = 1:5))

}
