% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{score_scale}
\alias{score_scale}
\alias{score_scale.data.frame}
\alias{score_scale.default}
\title{Scaling}
\usage{
score_scale(x, center = mean, scale = stats::sd)

\method{score_scale}{data.frame}(x, center = mean, scale = stats::sd)

\method{score_scale}{default}(x, center = mean, scale = stats::sd)
}
\arguments{
\item{x}{A vector of values}

\item{center}{A center}

\item{scale}{A scale}
}
\value{
A vector of scales scores
}
\description{
Scale scores with centrality and scale
}
\examples{
score_scale(1:10)
score_scale(1:10, center = 3, scale = 1)

x <- matrix(1:15, ncol = 3)
score_scale(x)
score_scale(as.data.frame(x))
}
