% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twenty-package.R
\docType{package}
\name{twenty-package}
\alias{twenty}
\alias{twenty-package}
\title{twenty: A Toolkit for Scores}
\description{
Provides functions for working with scores.
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
