% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/west.R
\docType{data}
\name{west}
\alias{west}
\title{west The original raw data}
\format{
A dataframe
}
\source{
https://github.com/ramhiser/datamicroarray/blob/master/data/west.RData
}
\usage{
data(west)
}
\description{
Data from Predicting the clinical status of human breast cancer by using gene expression profiles
}
\examples{
data(west)
}
\references{
Predicting the clinical status of human breast cancer by using gene expression profiles DOI: 10.1073/pnas.201162998
}
\keyword{datasets}
