% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/west_data_wide.R
\docType{data}
\name{west_data_wide}
\alias{west_data_wide}
\title{west_data_wide A wide dataset}
\format{
A dataframe
}
\source{
https://github.com/ramhiser/datamicroarray/blob/master/data/west.RData
}
\usage{
data(west_data_wide)
}
\description{
Data from Predicting the clinical status of human breast cancer by using gene expression profiles
}
\examples{
data(west_data_wide)
}
\references{
Predicting the clinical status of human breast cancer by using gene expression profiles DOI: 10.1073/pnas.201162998
}
\keyword{datasets}
