% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R
\name{anchorize}
\alias{anchorize}
\title{Anchors}
\usage{
anchorize(x, anchors = NULL, reverse = FALSE)
}
\arguments{
\item{x}{A vector of values}

\item{anchors}{A vector of anchors}

\item{reverse}{If \code{TRUE} reverses the scores in \code{x} \code{\link[=score_reverse]{score_reverse()}}}
}
\description{
Anchors
}
\examples{
x <- c(0.826, 0.674, 0.543, 0.105, 0.592)
anchors <- seq(-1, 1, .1)
anchorize(x, anchors)

}
