% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_long.R
\name{convert_to_long}
\alias{convert_to_long}
\title{Table of Genes in a long format}
\usage{
convert_to_long(the_dataset)
}
\arguments{
\item{the_dataset}{This is a dataset often in a wide format.}
}
\value{
gene_data
}
\description{
Transformation of a gene expression matrix into a 3 column table.
}
\examples{
my_long_data <- convert_to_long(gene_expression_table)
my_long_data
}
\author{
Oriade Latifah Simpson
}
