% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_manhattan.R
\name{make_manhattan}
\alias{make_manhattan}
\title{Creation of a Manhattan Plot}
\usage{
make_manhattan(
  wrangled_dataset,
  dna_segment,
  negative_pvalue,
  significance_id,
  gene_name
)
}
\arguments{
\item{wrangled_dataset}{This is a dataset that has been cleaned}

\item{dna_segment}{This is the gene column}

\item{negative_pvalue}{This is the negative p value column}

\item{significance_id}{This is the identified_as which shows if the gene is significant or not.}

\item{gene_name}{This is the gene identification label.}
}
\value{
visualisation_of_genes
}
\description{
Creation of a Manhattan plot from a small selection of expressed genes.
}
\examples{
manhattan_plot <- make_manhattan(nested_dataset)
manhattan_plot
}
\author{
Oriade Latifah Simpson, \email{s172084@dtu.dk}
}
