% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rotation_table.R
\name{make_rotation_table}
\alias{make_rotation_table}
\title{Create a rotation table from a modelled PCA object.}
\usage{
make_rotation_table(pca_fit)
}
\arguments{
\item{pca_fit}{This is a model object.}
}
\value{
rt
}
\description{
Create a rotation table from a modelled PCA object.
}
\examples{
rot_tab <- make_rotation_table(fitted_pca_object)
rot_tab
}
\author{
Miss Oriade Latifah Simpson, \email{s172084@dtu.dk}
}
