% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_pca.R
\name{perform_pca}
\alias{perform_pca}
\title{Example dataset Model data using Principal Components analysis}
\usage{
perform_pca(provided_dataset)
}
\arguments{
\item{provided_dataset}{This is the dataset that is procided to the function}
}
\value{
modelled_data
}
\description{
Perform data modelling using principal components analysis
}
\examples{
pca_fit <- perform_pca(a_dataset)
pca_fit
}
\author{
Miss Oriade Latifah Simpson, \email{s172084@dtu.dk}
}
